ccextractor: Add force-expose-caption-pad property
When this property is enabled, ccextractor will expose caption pad
on the first buffer even if caption meta is not attached in the buffer.

This feature can be useful when caption branch needs to be
strictly synchronized with video branch via pushed GAP event.

For instance, application might want to produce text related data
even if it doesn't exist (as a dummy or similar form).
Live encoding would be a likely scenario. But by the nature of text,
stream might not contain any caption data at the beginning then
ccextractor cannot notify downstream of emptiness via GAP event
since caption pad is not added yet.